end cycle + cleaning
