+++ python/mainPath.py
@@ -4,8 +4,30 @@ from domain.Edge import Edge
 from domain.PathProblem import PathProblem
 from parsing.readFromResources import path
 from shortestPath import shortestPath
+
+from sys import argv
 import random
 from functools import lru_cache
+
+
+def mainPath() -> None:
+    try:
+        random.seed(float(argv[1]))
+    except Exception:
+        pass
+
+    printResult("grid", path("grid.txt"))
+    printResult("spiral", path("spiral.txt"))
+    printResult("bunker", path("bunker.txt"))
+    printResult("line", path("line.txt"))
+    printResult("Small hole", path("small_hole.txt"))
+    printResult("Stick man", path("stickman.txt"))
+
+    printResult("Simple graph 5x4 grid", PathProblem(shuffle(gridNorthEast(5, 4)), Point(0, 0), Point(3, 3)), False)
+    printResult("Large graph 50x40 grid", PathProblem(shuffle(gridNorthEast(50, 40)), Point(0, 0), Point(30, 30)), False)
+    printResult("Large graph  with cycle 50x40 grid", PathProblem(shuffle(gridFull(50, 40)), Point(0, 0), Point(30, 30)))
+    # too long yet for mapping in printPath
+    # printResult("Huge graph 200x300 grid", PathProblem(shuffle(gridNorthEast(200, 300)), Point(0, 0), Point(30, 30)), False)
 
 
 @lru_cache()
@@ -28,6 +50,22 @@ def gridFull(width: int, height: int) -> "list[Edge]":
             for elem in elemList]
 
 
+def shuffle(grid: list) -> list:
+    random.shuffle(grid)
+    return grid
+
+
+def printResult(title: str, problem: PathProblem, isGrid: bool = True) -> None:
+    print(f"{title}\nfrom {problem.start} to {problem.end}")
+    result: Path = shortestPath(problem, isGrid)
+    if result.length():
+        print(result.length())
+        printPath(problem, result)
+    else:
+        print("no path found")
+    print("\n\n")
+
+
 def printPath(problem: PathProblem, path: Path) -> None:
     points: "iter[Point]" = sum(map(lambda edge: (edge.to, edge.from_), problem.graph), ())
 
@@ -48,37 +86,6 @@ def printPath(problem: PathProblem, path: Path) -> None:
             else:
                 print("#", end="")
         print()
-
-
-def printResult(title: str, problem: PathProblem, isGrid: bool = True) -> None:
-    print(f"{title}\nfrom {problem.start} to {problem.end}")
-    result: Path = shortestPath(problem, isGrid)
-    if result.length():
-        print(result.length())
-        printPath(problem, result)
-    else:
-        print("no path found")
-    print("\n\n")
-
-
-def mainPath() -> None:
-    printResult("grid", path("grid.txt"))
-    printResult("spiral", path("spiral.txt"))
-    printResult("bunker", path("bunker.txt"))
-    printResult("line", path("line.txt"))
-    printResult("Small hole", path("small_hole.txt"))
-    printResult("Stick man", path("stickman.txt"))
-
-    printResult("Simple graph 5x4 grid", PathProblem(shuffle(gridNorthEast(5, 4)), Point(0, 0), Point(3, 3)), False)
-    printResult("Large graph 50x40 grid", PathProblem(shuffle(gridNorthEast(50, 40)), Point(0, 0), Point(30, 30)), False)
-    printResult("Large graph  with cycle 50x40 grid", PathProblem(shuffle(gridFull(50, 40)), Point(0, 0), Point(30, 30)))
-    # too long yet for mapping in printPath
-    # printResult("Huge graph 200x300 grid", PathProblem(shuffle(gridNorthEast(200, 300)), Point(0, 0), Point(30, 30)), False)
-
-
-def shuffle(grid: list) -> list:
-    random.shuffle(grid)
-    return grid
 
 
 if __name__ == "__main__":
